[🛡️ Safety] Pre-run snapshot for WP-043
Auto-committed by Levitate before starting Work Package execution.
This commit preserves the exact filesystem state that existed when the run began.

- Work Package: WP-043
- Trigger: Auto-Safety Init

🤖 Automated checkpoint by Levitate AI
⏰ 2026-02-09 23:58:28
📋 Run: SAFETY-WP-043

+++ crates/fibonacci_math/src/fibonacci_heap.rs
@@ -291,7 +291,7 @@ impl FibonacciHeap {
             // While there's another node with the same degree
             while let Some(y) = degree_array[d] {
                 // Make sure x is the smaller key
-                if (self.nodes[x].key as f64) > (self.nodes[y].key as f64) { // Explicit f64 comparison
+                if self.nodes[x].key > self.nodes[y].key { // Type annotation needed for comparison
                     std::mem::swap(&mut x, &mut y);
                 }
